Fuck My Life

+++ database/migrations/2024_11_26_174058_create_products_table.php
@@ -17,7 +17,7 @@ return new class extends Migration
             $table->id();
             $table->string('name');
             $table->string('description');
-            $table->integer('price');
+            $table->decimal('price',10, 2);
             $table->integer('stock');
             $table->foreignIdFor(Store::class)->constrained()->cascadeOnDelete();
             $table->timestamps();
